12/9/2025 commit

+++ src/main/java/com/example/Dynamo_Backend/entities/TempProcess.java
@@ -24,7 +24,7 @@ public class TempProcess {
     @Column(name = "machine_id", nullable = false)
     private Integer machineId;
     @Column(name = "pg_time", nullable = false)
-    private Long pgTime;
+    private Float pgTime;
     @Column(name = "manufacturing_point", nullable = false)
     private Integer point;
 }
