+++ src/main/java/com/example/Dynamo_Backend/service/implementation/OperateHistoryImplementation.java
@@ -87,7 +87,7 @@ public class OperateHistoryImplementation implements OperateHistoryService {
 
                                                 TempProcess tempProcess = tempProcessRepository
                                                                 .findByProcessId(drawingCodeProcessDto.getProcessId());
-                                                Long pgTime = operateHistory.getPgTime()
+                                                Float pgTime = operateHistory.getPgTime()
                                                                 - tempProcess.getPgTime();
                                                 Integer point = operateHistory.getManufacturingPoint()
                                                                 - tempProcess.getPoint();
